requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
pdfplumber
pytesseract
pdf2image
Pillow
